Add ruby version of mate.
This is not saving back (on the server) but otherwise works. It takes one or more files as argument and opens this in the local version of Avian.

I have this in my .ssh/config for the host:

    RemoteForward 25757 localhost:25757
    ForwardAgent yes
    ControlPath ~/.ssh/master-%r@%h:%p
    ControlMaster auto

The first line is the essential one, the next ones are so that it re-uses the connection, that way, it won’t try to setup forward for same port if a second session is opened.

The protocol with Avian is very simply, it is: First line is just «command» and then zero or more «variable»: «value» to provide settings.

One special case is that the variable ‘data’ has a integer as value which tells how many raw bytes should be read next, after those, more variables may follow (even another data variable, in which case the bytes read should be appended to those already read, this is to allow transferring data in chunks, e.g. TM may not know in advance how many bytes a file will amount to (due to encodings and such)).

The rmate app tells Avian that it wants the file data back on save with ‘data-on-save: YES’ (look in the source). Avian sends back ‘save’ as command and then the data via the variable system explained above. It sends back the ‘close’ command when a file is closed. It puts the token among the variables sent back if one was provided during open.