Fix rmate saving.
FileUtils.cp throws an exception on failure (rather than false return), as will a failed File.open, so this allows us to catch errors in a simple rescue block.
FileUtils.cp will try to overwrite an existing file at the target path, so there is no need to explicitly unlink an existing backup first.

Long-term we probably want some way for rmate to notify the server that the save failed.