Alter shebang and add encoding comment.
We use ‘env’ to find ‘ruby’ so that it is still found if installed outside ‘/usr/bin’ (presuming its location is listed in the PATH environment variable).

Since not all implementations of ‘env’ support additional arguments to the executable we wish to run, there now is an encoding comment to replace the -KU argument (for ruby 1.9).